+++ src/hooks/addressHooks.ts
@@ -1,5 +1,5 @@
+import { fetchAddresses } from '@/services/address';
 import { Address } from '@/types/userAddressType';
-import { fetchAddresses } from '@/app/api/mypage/address/list/route';
 import { useQuery } from '@tanstack/react-query';
 
 export const useFetchAddresses = (userId: string | null) => {
